refactor: move team tests to own module
